enhance variable names and mapping between maze grid and canvas graphics

+++ src/main/java/org/example/maze/Coordinate.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class Coordinate {
-    private int row, col;
+    private double row, col;
 }
